Refactor Agents.py into multiple classes
